get and set tests moved from ConstructorTest to Test, files BossMonster and BossMonsterTest created, tests created for BossMonster constructor, tests all working
