+++ test/com.interoligt.rougelike/Main/BasicMonsterTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 
 class BasicMonsterTest {
-BasicMonster monster;
+    BasicMonster monster;
 
     @Test
     void isAliveTest(){
@@ -34,5 +34,45 @@ BasicMonster monster;
         assertFalse(monster.isAlive());
         monster.setAlive(true);
         assertTrue(monster.isAlive());
+    }
+
+    @Test
+    void toStringTest(){
+        assertEquals(monster.toString(),"Name: Spider, Level: 5, Element: Fire");
+    }
+
+    @Test
+    void getNameTest(){
+        assertEquals(monster.getName(),"Spider");
+    }
+
+    @Test
+    void getLevelTest(){
+        assertEquals(monster.getLevel(),5);
+    }
+
+    @Test
+    void getElementTest(){
+        assertEquals(monster.getElement(),Element.FIRE);
+    }
+
+    @Test
+    void getBaseHealthTest(){
+        assertEquals(monster.getBaseHealth(), 50);
+    }
+
+    @Test
+    void getBaseArmourTest(){
+        assertEquals(monster.getBaseArmour(),8);
+    }
+
+    @Test
+    void getBaseDamageTest(){
+        assertEquals(monster.getBaseDamage(),4);
+    }
+
+    @Test
+    void getBaseSpeedTest(){
+        assertEquals(monster.getBaseSpeed(),2);
     }
 }
